Add access to S3 path validator
This validator checks the Lambda has permission to perform GetObject on the requested object.
The app grants permission to download S3 objects via presigned URLs.
The presigned URLs are signed by the credentials granted to the Lambda.
For the presigned URLs to work the Lambda credentials have to have access to perform GetObject even though the Lambda code doesn't perform GetObject directly.