Add test for CORS supported browsers.
Browsers with JS disabled should work.
For non-CORS JS enabled browsers we should now fall back on the no JS version with an alert.
For older IE we have added a function to check if console exists before console logging.

This should fix Windows 7 IE9.
It does not yet work for Windows 8 IE10 which claims CORS support but isn't working.