Limit the number of files that can be displayed
It appears as though the response body cannot be larger that 1MB when
using an ALB in front of a Lambda[1].

Users are currently seeing broken journeys so I'm making this change
without tests to get it out before the weekend. I'm going to raise a
separate story to fix this properly.

[1] https://docs.aws.amazon.com/elasticloadbalancing/latest/application/load-balancer-troubleshooting.html#http-502-issues